+++ demisto_sdk/commands/content_graph/strict_objects/common.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from typing import Any, Optional, Sequence
+from typing import Any, Dict, Optional, Sequence
 
 import pydantic
 from pydantic import BaseModel, Extra, validator
@@ -47,6 +47,22 @@ class BaseStrictModel(BaseModel, ABC):
             "propagation_labels",
             "sort_values",
             "playbook_id",
+            "query",
+            "playbook_input_query",
+            "suppression_duration",  # correlation rules
+            "suppression_fields",  # correlation rules
+            "user_defined_category",  # correlation rules
+            "user_defined_severity",  # correlation rules
+            "investigation_query_link",  # correlation rules
+            "cron_tab",  # correlation rules
+            "search_window",  # correlation rules
+            "sort",  # widget
+            "params",  # widget
+            "cache",  # widget
+            "tags",  # modeling rule
+            "to_value",  # report
+            "from_value",  # report
+            "description",  # xsiam_dashboard
         }:
             # The assertion is caught by pydantic and converted to a pydantic.ValidationError
             assert value is not None, f"{value} may not be None"
@@ -127,4 +143,86 @@ ID_DYNAMIC_MODEL = create_dynamic_model(
     type_=Optional[Any],
     default=None,
     include_without_suffix=True,
+)
+KEY_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="key",
+    type_=Optional[str],
+    default=None,
+)
+VALUE_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="value",
+    type_=Optional[Any],
+    default=None,
+)
+PLAYBOOK_INPUT_QUERY_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="playbookInputQuery",
+    type_=Optional[Any],
+    default=None,
+)
+FORM_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="form",
+    type_=Optional[Dict],
+    default=None,
+    include_without_suffix=True,
+)
+MESSAGE_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="message",
+    type_=Optional[Dict],
+    default=None,
+    include_without_suffix=True,
+)
+SCRIPT_ARGUMENTS_LOWER_CASE_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="scriptarguments",
+    type_=Optional[Dict],
+    default=None,
+    include_without_suffix=True,
+)
+SCRIPT_ARGUMENTS_UPPER_CASE_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="scriptArguments",
+    type_=Optional[Dict],
+    default=None,
+    include_without_suffix=True,
+)
+SCRIPT_ID_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="scriptId",
+    type_=Optional[str],
+    default=None,
+    include_without_suffix=True,
+)
+IS_CONTEXT_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="iscontext",
+    type_=Optional[bool],
+    default=None,
+    include_without_suffix=True,
+)
+
+
+class _LeftOrRight(BaseStrictModel):
+    value: Any  # VALUE_DYNAMIC_MODEL doesn't have the raw 'value', only its variations
+
+
+LeftOrRight = create_model(
+    model_name="LeftOrRight",
+    base_models=(_LeftOrRight, VALUE_DYNAMIC_MODEL, IS_CONTEXT_DYNAMIC_MODEL),
+)
+
+LEFT_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="left",
+    type_=Optional[LeftOrRight],
+    default=None,
+    include_without_suffix=True,
+)
+
+RIGHT_DYNAMIC_MODEL = create_dynamic_model(
+    field_name="right",
+    type_=Optional[LeftOrRight],
+    default=None,
+    include_without_suffix=True,
+)
+
+SUFFIXED_ID_DYNAMIC_MODEL = create_dynamic_model(
+    # creating here with include_without_suffix == False
+    field_name="id",
+    type_=Optional[str],
+    default=None,
 )