+++ demisto_sdk/commands/validate/validators/ST_validators/ST110_is_valid_scheme.py
@@ -1,12 +1,32 @@
 from typing import Iterable, List, Union
 
 from demisto_sdk.commands.content_graph.objects import (
+    Classifier,
+    CorrelationRule,
+    Dashboard,
+    GenericDefinition,
+    GenericField,
+    GenericModule,
     GenericType,
     IncidentField,
     IncidentType,
     IndicatorField,
+    Job,
+    Layout,
+    LayoutRule,
+    Mapper,
+    ModelingRule,
+    ParsingRule,
+    Playbook,
+    PreProcessRule,
+    Report,
+    Widget,
+    Wizard,
+    XDRCTemplate,
+    XSIAMDashboard,
 )
 from demisto_sdk.commands.content_graph.objects.integration import Integration
+from demisto_sdk.commands.content_graph.objects.list import List as ListObject
 from demisto_sdk.commands.content_graph.objects.script import Script
 from demisto_sdk.commands.validate.validators.base_validator import (
     BaseValidator,
@@ -14,7 +34,32 @@ from demisto_sdk.commands.validate.validators.base_validator import (
 )
 
 ContentTypes = Union[
-    Integration, Script, IncidentField, IndicatorField, IncidentType, GenericType
+    Integration,
+    Script,
+    IncidentField,
+    IndicatorField,
+    IncidentType,
+    GenericType,
+    Classifier,
+    Layout,
+    LayoutRule,
+    Playbook,
+    CorrelationRule,
+    Dashboard,
+    GenericDefinition,
+    GenericField,
+    GenericModule,
+    Job,
+    ListObject,
+    Mapper,
+    ModelingRule,
+    ParsingRule,
+    PreProcessRule,
+    Report,
+    Widget,
+    Wizard,
+    XDRCTemplate,
+    XSIAMDashboard,
 ]